+++ src/pages/Dashboard/SideBar/index.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { FiBarChart2 } from 'react-icons/fi';
 
 import Dropdown from './Dropdown';
-import { dashboardNavigation } from '../../../resources';
+import { sideBarNavigation } from '../../../resources';
 import { Container, AmountInvested } from './styles';
 import AccountBalance from './AccountBalance';
 
@@ -15,7 +15,7 @@ const SideBar: React.FC = () => {
         Ver total investido <FiBarChart2 size="1.6rem" />
       </AmountInvested>
       {/* o entries ele pega as 'enradas' do objeto e transforma em um array */}
-      {Object.entries(dashboardNavigation).map(([key, value]) => (
+      {Object.entries(sideBarNavigation).map(([key, value]) => (
         <Dropdown key={key} icon={value.icon} sectionName={key} />
       ))}
     </Container>
