+++ src/resources/sidebar-navigation.tsx
